Basic blocks for joins

+++ include/interfaces/iterator_with_model.h
@@ -16,10 +16,10 @@ public:
   T key() { return iterator_->key(); }
   uint64_t current_pos() const { return iterator_->current_pos(); }
   uint64_t num_keys() const { return iterator_->num_keys(); }
-  double guessPositionMonotone(T target_key) {
+  uint64_t guessPositionMonotone(T target_key) {
     return model_->guessPositionMonotone(target_key);
   };
-  double guessPosition(T target_key) {
+  uint64_t guessPosition(T target_key) {
     return model_->guessPosition(target_key);
   };
   uint64_t bulkReadAndForward(uint64_t num_keys, char **data, uint64_t *len) {
